Set default values on Virtus attributes
When there are no explicit default values for Virtus attributes the way is
implemented means that there is an exception raised for flow control and many
Strings get allocated as part of the error message.

When using profiling tools the number of objects allocated when hitting the
prisoner step most of them where Strings caused by a lack of default values in
attribtues.

This is bad for various reasons:

- Many object allocations make a fragemented heap more likely which means that
  memory might not be possible for the be freed to the OS.
- Many object allocated means more work for the GC.
- Error handling in Ruby is
  [slow](http://blog.honeybadger.io/benchmarking-exceptions-in-ruby-yep-theyre-slow/)